refactor: extract quiz submission utils from QuizRenderer for Fast Refresh Move makeQuizSubmission, buildFillInTheBlankIdentityString, and makeFillInTheBlankSubmission to quizSubmissionUtils.ts so QuizRenderer only exports the component and Fast Refresh works correctly.
